Crate reth_stages

Staged syncing primitives for reth.

This crate contains the syncing primitives Pipeline and Stage, as well as all stages that reth uses to sync.

A pipeline can be configured using Pipeline::builder().

For ease of use, this crate also exposes a set of StageSets, which are collections of stages that perform specific functions during sync. Stage sets can be customized; it is possible to add, disable and replace stages in the set.

Examples

// Create a pipeline that can fully sync
Pipeline::builder()
    .with_tip_sender(tip_tx)
    .add_stages(DefaultStages::new(
        provider_factory.clone(),
        HeaderSyncMode::Tip(tip_rx),
        consensus,
        headers_downloader,
        bodies_downloader,
        executor_factory,
    ))
    .build(provider_factory);

Feature Flags

test-utils: Export utilities for testing

Modules

prelude

A re-export of common structs and traits.

sets

Built-in StageSets.

stages

Implementations of stages.

Structs

ExecInput

Stage execution input, see Stage::execute.

ExecOutput

The output of a stage execution.

MetricsListener

Metrics routine that listens to new metric events on the events_rx receiver. Upon receiving new event, related metrics are updated.

Pipeline

A staged sync pipeline.

PipelineBuilder

Builds a Pipeline.

PipelineStagesProgress

Pipeline stages progress.

StageSetBuilder

Helper to create and configure a StageSet.

UnwindInput

Stage unwind input, see Stage::unwind.

UnwindOutput

The output of a stage unwinding.

Enums

BlockErrorKind

Represents the specific error type within a block error.

ControlFlow

Determines the control flow during pipeline execution.

MetricEvent

Collection of metric events.

PipelineError

A pipeline execution error.

PipelineEvent

An event emitted by a Pipeline.

StageError

A stage execution error.

Traits

Stage

A stage is a segmented part of the syncing process of the node.

StageExt

Stage trait extension.

StageSet

Combines multiple Stages into a single unit.

Type Aliases

MetricEventsSender

Alias type for metric producers to use.

PipelineFut

The future that returns the owned pipeline and the result of the pipeline run. See Pipeline::run_as_fut.

PipelineWithResult

The pipeline type itself with the result of Pipeline::run_as_fut