Trait reth_stages::prelude::StageSet

pub trait StageSet<DB: Database>: Sized {
    // Required method
    fn builder(self) -> StageSetBuilder<DB>;

    // Provided method
    fn set<S: Stage<DB> + 'static>(self, stage: S) -> StageSetBuilder<DB> { ... }
}

Combines multiple Stages into a single unit.

A StageSet is a logical chunk of stages that depend on each other. It is up to the individual stage sets to determine what kind of configuration they expose.

Individual stages in the set can be added, removed and overridden using StageSetBuilder.

Required Methods

fn builder(self) -> StageSetBuilder<DB>

Configures the stages in the set.

Provided Methods

fn set<S: Stage<DB> + 'static>(self, stage: S) -> StageSetBuilder<DB>

Overrides the given Stage, if it is in this set.

Panics

Panics if the Stage is not in this set.

Object Safety

This trait is not object safe.