Module reth_stages::sets

Built-in StageSets.

The easiest set to use is DefaultStages, which provides all stages required to run an instance of reth.

It is also possible to run parts of reth standalone given the required data is present in the environment, such as ExecutionStages or HashingStages.

Examples

// Build a pipeline with all offline stages.

// Build a pipeline with all offline stages and a custom stage at the end.
Pipeline::builder()
    .add_stages(
        OfflineStages::new(executor_factory).builder().add_stage(MyCustomStage)
    )
    .build();

Structs

DefaultStages

A set containing all stages to run a fully syncing instance of reth.

ExecutionStages

A set containing all stages that are required to execute pre-existing block data.

HashingStages

A set containing all stages that hash account state.

HistoryIndexingStages

A set containing all stages that do additional indexing for historical state.

OfflineStages

A set containing all stages that do not require network access.

OnlineStages

A set containing all stages that require network access by default.