Module reth_stages::stages

Implementations of stages.

Structs

AccountHashingStage

Account hashing stage hashes plain account. This is preparation before generating intermediate hashes and calculating Merkle tree root.

BodyStage

The body stage downloads block bodies.

ExecutionStage

The execution stage executes all transactions and update history indexes.

ExecutionStageThresholds

The thresholds at which the execution stage writes state changes to the database.

FinishStage

The finish stage.

HeaderStage

The headers stage.

IndexAccountHistoryStage

Stage is indexing history the account changesets generated in ExecutionStage. For more information on index sharding take a look at reth_db::tables::AccountHistory

IndexStorageHistoryStage

Stage is indexing history the account changesets generated in ExecutionStage. For more information on index sharding take a look at reth_db::tables::StorageHistory.

SeedOpts

SeedOpts provides configuration parameters for calling AccountHashingStage::seed in unit tests or benchmarks to generate an initial database state for running the stage.

SenderRecoveryStage

The sender recovery stage iterates over existing transactions, recovers the transaction signer and stores them in TxSenders table.

StorageHashingStage

Storage hashing stage hashes plain storage. This is preparation before generating intermediate hashes and calculating Merkle tree root.

TotalDifficultyStage

The total difficulty stage.

TransactionLookupStage

The transaction lookup stage.

Enums

MerkleStage

The merkle hashing stage uses input from AccountHashingStage and StorageHashingStage to calculate intermediate hashes and state roots.

Constants

MERKLE_STAGE_DEFAULT_CLEAN_THRESHOLD

The default threshold (in number of blocks) for switching from incremental trie building of changes to whole rebuild.